Move `StateSnapshot` into Redwood
